refactor(mcp-server): move flat src into aura_mcp package (namespace pilot)
Pilot for the app namespace refactor. mcp-server's flat src/ layout (top-level
main/receptor/translator/effector) meant its wheel shipped generic modules
that collide with other services' `main`/`config` in a shared venv, and forced
a flat-src hatch hack to build at all.

- Move src/*.py + manifest.yaml into src/aura_mcp/ (git-tracked renames).
- Internal imports -> relative (`from .translator import ...`); external
  core imports (config, hive.cortex) unchanged.
- Tests import `from aura_mcp.<mod>`.
- pyproject: entry point -> `aura_mcp.main:main`; wheel target ->
  `packages = ["src/aura_mcp"]` (standard src-layout, drops the flat hack).

The built wheel now ships only `aura_mcp/*` — no top-level modules — so the
console-script/namespace collision is gone for this service. Makefile MCP_PATH
already puts src/ on the path, so `import aura_mcp` resolves with no infra
change.

Verified: `make test` green (core 114, api-gateway 1, telegram-bot 6,
mcp-server 14); `uv build --package aura-mcp --wheel` ships aura_mcp/ only.

Follow-ups (separate PRs): bee-keeper, then the deployed services
(core, api-gateway, telegram-bot) with their Dockerfiles.
