+++ synapses/mcp-server/src/aura_mcp/manifest.yaml
