+++ synapses/mcp-server/src/aura_mcp/translator.py
